pyproject.toml
[project]
name = "django-starter-core"
version = "0.1.0"
description = "Core library for DjangoStarter (Web) and DjangoStarter API"
authors = [
    {name = "DealiAxy", email = "dealiaxy@gmail.com"},
]
dependencies = [
    "django>=6.0",
    "django-ninja>=1.5.3",
    "pydantic>=2.0",
    "django-simple-history>=3.0",
    "django-simple-captcha>=0.6.3",
    "django-ratelimit>=4.1.0",
    "pyjwt>=2.0",
    "prometheus-client>=0.17",
    "redis>=7.0.0",
    "anyio>=4.0.0",
    "jinja2>=3.1",
    "orjson>=3.9",
    "faker>=20.0.0",
    "markdown>=3.7",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
pythonpath = ["src", "."]
norecursedirs = [".venv", "dist", "build"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
